rename publicapi pkg to public/api
