+++ viz/cmd/dashboard.go
@@ -8,7 +8,7 @@ import (
 
 	"github.com/linkerd/linkerd2/pkg/healthcheck"
 	"github.com/linkerd/linkerd2/pkg/k8s"
-	publicAPI "github.com/linkerd/linkerd2/pkg/publicapi"
+	api "github.com/linkerd/linkerd2/pkg/public"
 	"github.com/pkg/browser"
 	"github.com/spf13/cobra"
 )
@@ -82,7 +82,7 @@ func newCmdDashboard() *cobra.Command {
 			}
 
 			// ensure we can connect to the public API before starting the proxy
-			publicAPI.CheckPublicAPIClientOrRetryOrExit(healthcheck.Options{
+			api.CheckPublicAPIClientOrRetryOrExit(healthcheck.Options{
 				ControlPlaneNamespace: controlPlaneNamespace,
 				KubeConfig:            kubeconfigPath,
 				Impersonate:           impersonate,
